Update CPGen.py
First thing changed is a fix to the class scale, I forgot the ‘self’ in the initialisation of the class - classic.

first output = [55, 60, 55, 57, 55, 55, 57, 57, 55, 57, 55, 62, 60]
second output = [81, 81, 76, 81, 79, 79, 81, 81, 79, 81, 79, 79, 81] 

These two outputs show an inherent issue, which is that the order of list of possible notes determines what solution will be chosen. The first prioritised the first in the list, which were on the lower end, and this resulted in a melody mostly below middle c. The second on the other hand prioritised the last note in the list, and ended up mostly within the same 5th in the upper soprano range.